Refactored code && removed edition command
