+++ src/api/modules/index.ts
@@ -3,3 +3,4 @@ export * from './auth'
 export * from './build'
 export * from './prompts'
 export * from './utils'
+export * from './sponsor'
