+++ src/api/modules/sponsor/index.ts
@@ -1,0 +1,2 @@
+export * from './getEdition'
+export { default as setEdition } from './setEdition'
