+++ src/api/modules/sponsor/setEdition.ts
@@ -1,11 +1,11 @@
 import chalk from 'chalk'
-import { CommandError } from '../../api/error/errors'
-import { Sponsor } from '../../api/clients/IOClients/apps/Sponsor'
-import { SessionManager } from '../../api/session/SessionManager'
-import log from '../../api/logger'
-import { promptWorkspaceMaster } from '../../api/modules/utils'
-import { returnToPreviousAccount, switchAccount } from '../../api/modules/auth/switch'
-import { promptConfirm } from '../../api/modules/prompts'
+import { CommandError } from '../../error/errors'
+import { Sponsor } from '../../clients/IOClients/apps/Sponsor'
+import { SessionManager } from '../../session/SessionManager'
+import log from '../../logger'
+import { promptWorkspaceMaster } from '../../modules/utils'
+import { returnToPreviousAccount, switchAccount } from '../../modules/auth/switch'
+import { promptConfirm } from '../../modules/prompts'
 
 const promptSwitchToAccount = async (account: string, initial: boolean) => {
   const reason = initial
